fix: cyclic dependency between app.ts and server.ts is eliminated
